Tasks 2 and 3 Added FinalCulator class to calculate fines for overdue library books
Added ReturnBook function to allow books to be returned, this inclueds the ReturnBook end point iun the BookStockController, ReturnBook function in the OnLoanSerivce and ReturnBook function in the BookStockRepository.
I have also added a new table called Fines which store all fines against each borrower. There is a repository and controller for fines.
I have also added a GinesCalculator class to work out fines for books that are returned late.

There is also a new book in the SeedData class for testing books that are returned late